テーブル修正

+++ src/database/seeders/Traits/Insert.php
@@ -3,10 +3,12 @@
 namespace Database\Seeders\Traits;
 
 trait Insert {
-    private static function insertArray(array $columns, array $data_list, $model)
+    private static function insertArray(array $columns, array $data_list, $model, bool $is_query_delete = true)
     {
-        // マスタデータをリセットするためにデータを一括削除
-        $model->query()->delete();
+        if ($is_query_delete) {
+            // マスタデータをリセットするためにデータを一括削除
+            $model->query()->delete();
+        }
 
         // マスタデータを挿入
         foreach ($data_list as $data) {
